+++ src/index.html
@@ -19,7 +19,7 @@
             <form id="update-list">
                 <input id="add-to-list" type="text" placeholder="Add your list...">
                 <button class="add-btn">
-                    <span class="material-icons">&#xe5d9;</span>
+                    <span class="material-icons">&#xe145;</span>
                 </button>
             </form>
         </div>
