+++ src/style.css
@@ -14,7 +14,7 @@ body {
 }
 
 .todo-header {
-  padding: 5px 30px;
+  padding: 5px 35px;
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -94,6 +94,7 @@ input.checkbox {
 .material-icons {
   color: #b8b8b8;
   cursor: pointer;
+  font-size: 20px;
 }
 
 .edit {
@@ -104,7 +105,6 @@ input.checkbox {
 .delete {
   position: relative;
   right: 25px;
-  display: none;
 }
 
 .edit-mode {
